Title: 삼각형 만들기, Time: 96ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240321/삼각형 만들기/create-triangle.java
@@ -38,7 +38,7 @@ public class Main {
         
                     
                      ){
-                        int result = Math.abs((x1*y2 + x2*y3 + x3*y1) -(x2*y1 + x3*y2 + x1*y3))/2;
+                        int result = (int)Math.abs((x1*y2 + x2*y3 + x3*y1) -(x2*y1 + x3*y2 + x1*y3))/2;
                         max = Math.max(max, result);
                     }
                 }
